mysql-connections/test-app/pyproject.toml
[project]
name = "test-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Victor Nascimento",email = "victooyooth@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mysql-connector-python (>=9.4.0,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
